/groupFix changed window refresh to group reload
