+++ frontend/src/components/groups/group.css
@@ -19,7 +19,7 @@
 }
 
 .groupStyle {
-  width: 1100px;
+  min-width: 1100px;
   min-height: 600px;
   outline-color: blue;
   border: #e1e1e1 4px solid;
